requirements.txt
openai
python-dotenv
playwright
beautifulsoup4
pytest
pytest-html
google-generativeai
# --- NEW: Library for Windows desktop automation ---
pywinauto